Move the root Queries/Mutations to manager.models.gql
* And implement role checks for ComputeSession-related queries
